Modify Gestion arguments

Modify Gestion arguments

Usage

IAM.editArgs_Gest(object, ...)

# S4 method for iamArgs
IAM.editArgs_Gest(
  object,
  active = NULL,
  control = NULL,
  target = NULL,
  espece = NULL,
  delay = NULL,
  type = NULL,
  update = NULL,
  bounds = NULL,
  tac = NULL,
  fbar = NULL,
  effSup = NULL,
  mfm = NULL
)

Arguments

object

a iamArgs-class object

...

selected argument

active

Activation of the module. TRUE or FALSE input, but accept 1 or 0 values (any numeric will be interpreted as logical)

control

option to choose if the management will control effort reduction by reducing the number of vessels or number of trips per vessels.

target

management aim to reduce effort to reach TAC or Fbar, or TAC then Fbar.

espece

species to target for management.

delay

option to delay the management for n years. If the delay is larger than the year in model, the management will have no effect.

type

option to choose if effort reduction is additive ("+") or multiplicative ("x").

update

option to choose is effort reduction is from t0 or t-1 iteration

bounds

bounds for target reach at each model iteration. num of length 2.

tac

TAC values for each model iteration. Can be NA.

fbar

Fbar values for each model iteration. Can be NA.

effSup

Maximum effort per fleet and year. To limit nbds.

mfm

Ponderation fleet/metier for effort reduction.

Author

Maxime Jaunatre

Examples

data(IAM_argum_1984)
IAM.args_scenar(IAM_argum_1984)
#> $active
#> [1] 0
#> 
#> $ALLscenario
#> [1] "Sc_saisine"
#> 
#> $SELECTscen
#> [1] 1
#> 

# Activate scenario "Sc_saisine"
IAM_argum_1984 <- IAM.editArgs_Gest(IAM_argum_1984, active= TRUE)
summary(IAM_argum_1984)
#> My Input (IAM argument) :
#> Simulation of 3 dynamic species, 19 static species and 7 fleet
#> Simulation start in 1984 and end in 1995 (12 steps)
#> 
#> =======================================================================================
#>   SR module  |               Stock Recruitment             |      Noise       | Proba |
#> ---------------------------------------------------------------------------------------
#>     Species  |    function  :  param A ; param B ; param C |  Type :    sd    |  Type |
#>    ARC (XSA) |          Mean 3.641e+07  0.00e+00  0.00e+00 |  Norm | 0.00e+00 |   .   |
#>    COR (XSA) |          Mean 1.893e+07  0.00e+00  0.00e+00 |  Norm | 0.00e+00 |   .   |
#>    DAR (SS3) | not activated 0.000e+00  0.00e+00  0.00e+00 |  Norm | 0.00e+00 |   .   |
#> ---------------------------------------------------------------------------------------
#> 
#> =============================================================================
#>                      Gestion Module                                         | 
#> -----------------------------------------------------------------------------
#>     Species  |   control  |    target | delay | typeG | update |   bounds   |
#>    ARC (XSA) | Nb vessels |       TAC |    2  |   +   |   yes  |   0 :    0 |
#> -----------------------------------------------------------------------------
#>    TAC : NA NA NA NA NA NA NA NA NA NA NA NA 
#>   FBAR : NA NA NA NA NA NA NA NA NA NA NA NA 
#> ----------------------------------------------------------------------------
#> 
#> ============================================================
#> Economic : PerscCalc = 0 ; dr = 0.000 | No replicates      |
#> ------------------------------------------------------------
#> 
#>  The Scenario module is not active.