'iamInput' objects creator

'iamInput' objects creator

Usage

IAM.input(fileIN, fileSPEC, fileSCEN, fileSTOCH, ...)

# S4 method for character,missing,missing,missing
IAM.input(
  fileIN,
  t_init,
  nbStep = 20,
  t_hist_max = t_init,
  desc = "My Input",
  folderFleet = NULL,
  Fq_i = NULL,
  iniFq_i = NULL,
  Fq_fmi = NULL,
  iniFq_fmi = NULL,
  FqLwt_i = NULL,
  iniFqLwt_i = NULL,
  FqLwt_fmi = NULL,
  iniFqLwt_fmi = NULL,
  FqDwt_i = NULL,
  iniFqDwt_i = NULL,
  FqDwt_fmi = NULL,
  iniFqDwt_fmi = NULL,
  Nt0s1q = NULL,
  Ni0q = NULL,
  iniNt0q = NULL,
  matwt = NULL,
  Fg_fmi = NULL,
  dg_fmi = NULL,
  verbose = FALSE,
  ...
)

# S4 method for character,character,missing,missing
IAM.input(fileIN, fileSPEC, desc = "My Input", ...)

# S4 method for character,character,character,missing
IAM.input(fileIN, fileSPEC, fileSCEN, desc = "My Input", ...)

# S4 method for character,character,missing,character
IAM.input(fileIN, fileSPEC, fileSTOCH, desc = "My Input", ...)

# S4 method for character,character,character,character
IAM.input(fileIN, fileSPEC, fileSCEN, fileSTOCH, desc = "My Input", ...)

Arguments

fileIN

.xls standard IAM input file

fileSPEC

deprecated argument.

fileSCEN

deprecated argument.

fileSTOCH

deprecated argument.

...

Further arguments used when the input is a .xls. Described below.

t_init

Only for first signature. Initial year.

nbStep

Only for first signature. Number of timesteps (including initial year).

t_hist_max

Only for first signature. Last year considered for 'historical' slot.

desc

Object descriptor (default value : "My Input").

folderFleet

Folder containing fleets input sheets (Optionnal. Default value : NULL).

Fq_i

List containing SS3 parameters arrays for quarterly stock dynamic simulation. One element per considered species, all names must match with input species. Fishing mortality per season, morph, and age. (Optionnal. Default value : NULL. Example : list(hake=array(...,dim=c(4,4,nAge))).

Fq_fmi

See above. Fishing mortality per season, morph, fleet, "economic" metiers and age (Optionnal. Default value : NULL).

Ni0q

See above. Recruits numbers per season (Optionnal. Default value : NULL).