Complete IAM dataset

Dataset anonymised, with 7 fleets and 3 dynamic species, one being with a SS3 dynamic. Time begin in 1984 and end in 1995

Usage

IAM_input_1984

IAM_argum_1984

Format

Two different part of parameter for a IAM.model simulation :

IAM_input_1984 - a iamInput-class object. Created from a .xlsx file. See Details.

IAM_argum_1984 - a iamArgs-class object. Created from the input, to modify Recruitment, Gestion and Economic parameters

An object of class iamArgs of length 1.

Source

Based on Florence Briton dataset.

Details

This dataset is created with IAM.input-methods function. This function load a .xlsx file, require a t_init and nbStep. If a SS3 species is included in the .xlsx file, then additionnal data must be provided.

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Examples

if (FALSE) {

# Code used to create this object from raw .xlsx file. More information in vignette.
load(IAM_example("IAM_SS3_1984.RData"))
IAM_input_1984 <- IAM::IAM.input(
  fileIN = IAM_example("inputFile.xlsx"),
  t_init = 1984, nbStep = 12, folderFleet = IAM_example("fleets"),
  Fq_i = list(DAR = iniFq_i), iniFq_i = list(DAR = iniFq_i),
  Fq_fmi = list(DAR = iniFq_fmi), iniFq_fmi = list(DAR = iniFq_fmi),
  FqLwt_i = list(DAR = iniFqLwt_i),iniFqLwt_i = list(DAR = iniFqLwt_i),
  FqLwt_fmi = list(DAR = iniFqLwt_fmi), iniFqLwt_fmi = list(DAR = iniFqLwt_fmi),
  FqDwt_i = list(DAR = iniFqDwt_i), iniFqDwt_i = list(DAR = iniFqDwt_i),
  FqDwt_fmi = list(DAR = iniFqDwt_fmi), iniFqDwt_fmi = list(DAR = iniFqDwt_fmi),
  Nt0s1q = list(DAR = Nt0s1q), Ni0q = list(DAR = Ni0q),
  iniNt0q = list(DAR = iniNt0q), matwt = list(DAR = mat_morphage)
)

IAM_argum_1984 <- IAM.input2args(IAM_input_1984)
IAM_argum_1984@arguments$Eco$dr <- 0.04
IAM_argum_1984@arguments$Eco$perscCalc <- as.integer(1)
}

data(IAM_input_1984)
data(IAM_argum_1984)